+++ game/exceptions.py
@@ -1,0 +1,8 @@
+class LimitedMove(Exception):
+    pass
+
+class InvalidMove(Exception):
+    pass
+
+class OutOfBoard(Exception):
+    pass