+++ game/pieces.py
@@ -1,3 +1,5 @@
+from game.exceptions import LimitedMove, InvalidMove, OutOfBoard
+
 
 class Pieces(): ###Color = black or white, Piece = Queen, King, Pawn, Rook, Bishop, Knight, Position = (row , column) ♜
     def __init__(self, piece, color, initial_position):
@@ -36,7 +38,7 @@ class Pieces(): ###Color = black or white, Piece = Queen, King, Pawn, Rook, Bish
             return True
     
     def diagonal(self, new_position):
-        i_position = self.__position__
+        i_position = [self.__position__[0], self.__position__[1]]
         i = 0
         range = new_position[0] - i_position[0]
         if i_position[0] != new_position[0] and i_position[1] != new_position[1]:
@@ -50,6 +52,56 @@ class Pieces(): ###Color = black or white, Piece = Queen, King, Pawn, Rook, Bish
                 return False
         else:
             return False
+        
+    def valid_or_invalid(self, new_position):
+        diagonal = self.diagonal(new_position)
+        straight = self.straight_line(new_position)
+        if diagonal == straight:
+            raise InvalidMove()
+        
+    def on_board(self, new_position):
+        x = new_position[0]
+        y = new_position[1]
+        if x < 0 or x > 7:
+            raise OutOfBoard()
+        elif y < 0 or y > 7:
+            raise OutOfBoard()
+
+class Queen(Pieces):
+    def movement(self, new_position):
+        try:
+            self.on_board(new_position)
+            self.valid_or_invalid(new_position)
+        except Exception as e:
+            print(e)
+            raise
+    
+class King(Pieces):
+    def limit(self, new_position):
+        row = self.__position__[0]
+        column = self.__position__[1]
+        list = [[row + 1, column + 1],
+                [row + 1, column],
+                [row + 1, column - 1],
+                [row - 1, column + 1],
+                [row - 1, column],
+                [row - 1, column - 1],
+                [row, column + 1],
+                [row, column - 1]]
+        
+        if new_position not in list:
+            raise LimitedMove()
+        
+    def movement(self, new_position):
+        try:
+            self.on_board(new_position)
+            self.limit(new_position)
+            self.valid_or_invalid(new_position)
+        except Exception as e:
+            print(e)
+            raise
+        
+