+++ CHANGELOG.md
@@ -27,4 +27,10 @@
 
 - Added straight_line and diagonal. Checks if the movement for a piece is valid
 
+- Class Queen and King now verify the movement intended to make
+
+# CREATED
+
+- File exception
+